Update LICENSE
Some of the source files are based on other open source works. The Merkle Distributor contract comes from Uniswap: https://github.com/Uniswap/merkle-distributor
It has a GPL-3.0 license on it, thus this project inherits that license.
As far as I'm concerned, license issues should be sorted out in the repository in it's current state.
If you happen to know otherwise, don't hesitate to contact us or open an issue and we will solve any remaining problems as soon as possible.